cleanup: Refactor ResumableUploadResponse.
In a future PR I will need to include
`storage/internal/resumable_upload_session.h` from a public header.
Refactoring `internal::ResumableUploadResponse` out of the bulky header
where it was will minimize the impact of that future change.
