+++ google/cloud/storage/internal/object_requests_test.cc
@@ -561,115 +561,6 @@ TEST(ObjectRequestsTest, QueryResumableUpload) {
   EXPECT_THAT(actual, HasSubstr(url));
 }
 
-TEST(ObjectRequestsTest, ResumableUploadResponse) {
-  auto actual = ResumableUploadResponse::FromHttpResponse(
-                    HttpResponse{200,
-                                 "test-payload",
-                                 {{"ignored-header", "value"},
-                                  {"location", "location-value"},
-                                  {"range", "bytes=0-2000"}}})
-                    .value();
-  EXPECT_EQ("test-payload", actual.payload);
-  EXPECT_EQ("location-value", actual.upload_session_url);
-  EXPECT_EQ(2000, actual.last_committed_byte);
-  EXPECT_EQ(ResumableUploadResponse::kDone, actual.upload_state);
-
-  std::ostringstream os;
-  os << actual;
-  auto actual_str = os.str();
-  EXPECT_THAT(actual_str, HasSubstr("upload_session_url=location-value"));
-  EXPECT_THAT(actual_str, HasSubstr("last_committed_byte=2000"));
-}
-
-TEST(ObjectRequestsTest, ResumableUploadResponseNoLocation) {
-  auto actual =
-      ResumableUploadResponse::FromHttpResponse(
-          HttpResponse{308, "test-payload", {{"range", "bytes=0-2000"}}})
-          .value();
-  EXPECT_EQ("test-payload", actual.payload);
-  EXPECT_EQ("", actual.upload_session_url);
-  EXPECT_EQ(2000, actual.last_committed_byte);
-  EXPECT_EQ(ResumableUploadResponse::kInProgress, actual.upload_state);
-}
-
-TEST(ObjectRequestsTest, ResumableUploadResponseNoRange) {
-  auto actual =
-      ResumableUploadResponse::FromHttpResponse(
-          HttpResponse{201, "test-payload", {{"location", "location-value"}}})
-          .value();
-  EXPECT_EQ("test-payload", actual.payload);
-  EXPECT_EQ("location-value", actual.upload_session_url);
-  EXPECT_EQ(0, actual.last_committed_byte);
-  EXPECT_EQ(ResumableUploadResponse::kDone, actual.upload_state);
-}
-
-TEST(ObjectRequestsTest, ResumableUploadResponseMissingBytesInRange) {
-  auto actual = ResumableUploadResponse::FromHttpResponse(
-                    HttpResponse{308,
-                                 "test-payload",
-                                 {{"location", "location-value"},
-                                  {"range", "units=0-2000"}}})
-                    .value();
-  EXPECT_EQ("test-payload", actual.payload);
-  EXPECT_EQ("location-value", actual.upload_session_url);
-  EXPECT_EQ(0, actual.last_committed_byte);
-  EXPECT_EQ(ResumableUploadResponse::kInProgress, actual.upload_state);
-}
-
-TEST(ObjectRequestsTest, ResumableUploadResponseMissingRangeEnd) {
-  auto actual = ResumableUploadResponse::FromHttpResponse(
-                    HttpResponse{308, "test-payload", {{"range", "bytes=0-"}}})
-                    .value();
-  EXPECT_EQ("test-payload", actual.payload);
-  EXPECT_EQ("", actual.upload_session_url);
-  EXPECT_EQ(0, actual.last_committed_byte);
-  EXPECT_EQ(ResumableUploadResponse::kInProgress, actual.upload_state);
-}
-
-TEST(ObjectRequestsTest, ResumableUploadResponseInvalidRangeEnd) {
-  auto actual =
-      ResumableUploadResponse::FromHttpResponse(
-          HttpResponse{308, "test-payload", {{"range", "bytes=0-abcd"}}})
-          .value();
-  EXPECT_EQ("test-payload", actual.payload);
-  EXPECT_EQ("", actual.upload_session_url);
-  EXPECT_EQ(0, actual.last_committed_byte);
-  EXPECT_EQ(ResumableUploadResponse::kInProgress, actual.upload_state);
-}
-
-TEST(ObjectRequestsTest, ResumableUploadResponseInvalidRangeBegin) {
-  auto actual =
-      ResumableUploadResponse::FromHttpResponse(
-          HttpResponse{308, "test-payload", {{"range", "bytes=abcd-2000"}}})
-          .value();
-  EXPECT_EQ("test-payload", actual.payload);
-  EXPECT_EQ("", actual.upload_session_url);
-  EXPECT_EQ(0, actual.last_committed_byte);
-  EXPECT_EQ(ResumableUploadResponse::kInProgress, actual.upload_state);
-}
-
-TEST(ObjectRequestsTest, ResumableUploadResponseUnexpectedRangeBegin) {
-  auto actual =
-      ResumableUploadResponse::FromHttpResponse(
-          HttpResponse{308, "test-payload", {{"range", "bytes=3000-2000"}}})
-          .value();
-  EXPECT_EQ("test-payload", actual.payload);
-  EXPECT_EQ("", actual.upload_session_url);
-  EXPECT_EQ(0, actual.last_committed_byte);
-  EXPECT_EQ(ResumableUploadResponse::kInProgress, actual.upload_state);
-}
-
-TEST(ObjectRequestsTest, ResumableUploadResponseNegativeEnd) {
-  auto actual =
-      ResumableUploadResponse::FromHttpResponse(
-          HttpResponse{308, "test-payload", {{"range", "bytes=0--7"}}})
-          .value();
-  EXPECT_EQ("test-payload", actual.payload);
-  EXPECT_EQ("", actual.upload_session_url);
-  EXPECT_EQ(0, actual.last_committed_byte);
-  EXPECT_EQ(ResumableUploadResponse::kInProgress, actual.upload_state);
-}
-
 ObjectMetadata CreateObjectMetadataForTest() {
   // This metadata object has some impossible combination of fields in it. The
   // goal is to fully test the parsing, not to simulate valid objects.
