+++ google/cloud/storage/internal/retry_resumable_upload_session.h
@@ -18,6 +18,7 @@
 #include "google/cloud/storage/internal/resumable_upload_session.h"
 #include "google/cloud/storage/retry_policy.h"
 #include "google/cloud/storage/version.h"
+#include <memory>
 
 namespace google {
 namespace cloud {
